pyproject.toml
[tool.poetry]
name = "tap-oenb-meldefonds"
version = "0.0.0"
description = "Singer.io tap for extracting tax reporting fund (i.e. Meldefond) data from the OeNB website."
authors = ["Algebrazebra <Algebrazebra@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
singer-python = "^5.12.2"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
cookiecutter = "^1.7.3"
pre-commit = "^2.17.0"
commitizen = "^2.21.2"

[tool.poetry.scripts]
tap-oenb-meldefonds = "tap_oenb_meldefonds.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
